share functionality fix
