+++ apps/frontend/app/(website)/studio/features/share/usePublicLinkSharing.tsx
@@ -1,5 +1,5 @@
+import { useShareLink } from "@features/share/useShareLink";
 import { createSharedLink } from "@studio/actions/createSharedLink";
-import { useShareLink } from "@studio/hooks/useShareLink";
 import { useCallback, useState } from "react";
 import toast from "react-hot-toast";
