+++ apps/frontend/app/(website)/studio/src/store/mod.ts
@@ -10,6 +10,7 @@ import { type RangeCommand, buildRanges } from "../utils/tree";
 import { INITIAL_STATE } from "./getInitialState";
 
 type ModStateValues = {
+  name: string;
   content: string | null;
   hasRuntimeErrors: boolean;
   parsedContent: TreeNode | null;
@@ -37,6 +38,7 @@ const getInitialState = (): ModStateValues => {
     : null;
 
   return {
+    name: "",
     content: INITIAL_STATE.codemodSource,
     hasRuntimeErrors: false,
     parsedContent,
@@ -49,6 +51,7 @@ const getInitialState = (): ModStateValues => {
 export const useModStore = create<ModState>(
   persist(
     (set, get) => ({
+      ...getInitialState(),
       setState: (newState) => set((state) => ({ ...state, ...newState })),
       setContent: (content) => {
         const parsed = parseSnippet(content);
@@ -67,11 +70,6 @@ export const useModStore = create<ModState>(
     }),
     {
       name: "mod-store",
-      merge: (persistedState, currentState) => ({
-        ...currentState,
-        ...persistedState,
-        ...(!persistedState && getInitialState()),
-      }),
     },
   ),
 );
