+++ apps/frontend/app/(website)/studio/src/store/snippets.ts
@@ -45,6 +45,7 @@ type SnippetsConfig = {
   currentType: EditorType;
   addPair: () => void;
   clearAll: () => void;
+  setInitialState: (state: Partial<SnippetsState>) => void;
   removePair: (index: number) => void;
   selectedPairIndex: number;
   engine: KnownEngines;
@@ -116,13 +117,13 @@ export const useSnippetsStore = create<SnippetsState>(
             ...get().editors,
             {
               name: `Test ${
-                get()
+                (get()
                   .getAllNames()
-                  .filter((name) => name.startsWith("Test "))
+                  .filter((name) => name.toLowerCase().startsWith("test "))
                   .map((name) => name.split(" ")[1])
                   .map(Number)
                   .filter(Boolean)
-                  .at(-1) + 1
+                  .at(-1) || 0) + 1
               }`,
               before: getSnippetInitialState(),
               after: getSnippetInitialState(),
@@ -221,6 +222,9 @@ export const useSnippetsStore = create<SnippetsState>(
             get().setSelection(index, editorType),
         };
       },
+      setInitialState: (state) => {
+        set(state);
+      },
       setEngine: (engine) =>
         set({
           engine,
@@ -260,12 +264,6 @@ export const useSnippetsStore = create<SnippetsState>(
     }),
     {
       name: "snippets-storage",
-      merge: (persistedState, currentState) => ({
-        ...currentState,
-        ...persistedState,
-        engine: persistedState.engine || INITIAL_STATE.engine,
-        editors: persistedState.editors || INITIAL_STATE.editors,
-      }),
     },
   ),
 );
